requirements.txt
streamlit==1.45.1
requests==2.32.3
python-dotenv==1.0.1
azure-storage-blob==12.25.1
